Split tests and a new one for handled error with 1 attachment
